moved api calls to api.js
